src/requirements.txt
boto3>=1.34.69,<2.0
botocore>=1.34.69,<2.0
pydantic>=2.6.3,<3.0
requests>=2.31.0,<3.0
twilio>=9.0.0,<10.0
PyNaCl>=1.5.0,<2.0
